FS20 needs parseFloat instead of parseInt
git-svn-id: https://fhem.svn.sourceforge.net/svnroot/fhem/trunk/fhem@2000 2b470e98-0d58-463d-a4d8-8e2adae1ed80

+++ www/pgm2/fhemweb.js
@@ -56,7 +56,7 @@ Slider(slider, min, stp, max, curr, cmd)
 {
   var sh = slider.firstChild;
   var lastX=-1, offX=0, maxX=0, val=-1;
-  min = parseInt(min); max = parseInt(max); stp = parseInt(stp);
+  min = parseFloat(min); max = parseFloat(max); stp = parseFloat(stp);
 
   function
   init()
